Update spanish.json
I just put some accent marks and corrected some ortography faults that I've found. Also, I've translated the sentence 'Mayor, due to threat of an earthquake, we are withdrawing disaster recovery funds'. Send me a private message on the forum or in Discord if I made something wrong. Sorry, I'm new in this.