Update spanish.json
I just put some accent marks and corrected some ortography faults that I've found. Also, I've translated the sentence 'Mayor, due to threat of an earthquake, we are withdrawing disaster recovery funds'. Send me a private message on the forum or in Discord if I made something wrong. Sorry, I'm new in this.

Former-commit-id: 05926abda75edc4ac6de5a9673f49362ee18f680 [formerly 94e2e6527345973fa900cbab354c16618a238b80]
Former-commit-id: df68021f49b3c0f1645049afd152e9a1ee0d51b6
Former-commit-id: ecb102a2570b0e2183135ec46eb72f5741b4fcc9