Resolution fix and minor changes on the vocabulary
The resolution and city in spanish is on female so instead of ''bajo'' and ''muy bajo'' it would be ''baja'' and ''muy baja'' and instead of ''pequeño'' and ''mediano'' it would be ''pequeña'' and ''mediana'', furthermore the adjective for resolution is ''alta'' and not ''grande'' although I must say it was perfectly understandable before the correction it more correct now. Some words seemed weird so I changed them :3.