Resolution fix and minor changes on the vocabulary
The resolution and city in spanish is on female so instead of ''bajo'' and ''muy bajo'' it would be ''baja'' and ''muy baja'' and instead of ''pequeño'' and ''mediano'' it would be ''pequeña'' and ''mediana'', furthermore the adjective for resolution is ''alta'' and not ''grande'' although I must say it was perfectly understandable before the correction it more correct now. Some words seemed weird so I changed them :3.

Former-commit-id: 24cc2eefcddb99377000ed244a76d541d3fa1eba [formerly 8a62175fe3a458c609f60c8066c8bbdf0af66227]
Former-commit-id: 74223e3e88c88d0386af22adf29a74e2d23a08e0
Former-commit-id: 8587b0fc7db675f453aaa86f082cd8eb6953fa83